+++ app/(tabs)/index.tsx
@@ -1,47 +1,13 @@
 import { Linking, StyleSheet, TouchableOpacity } from "react-native";
 import EditScreenInfo from "@/components/EditScreenInfo";
 import { Text, View } from "@/components/Themed";
-import React, { useState } from "react";
+import React from "react";
 
 
 export default function TabOneScreen() {
-  const [href, setHref] = useState("");
-  const fetchApi = async () => {
-    const response = await fetch(
-      "https://places.googleapis.com/v1/places:searchText",
-      {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          "X-Goog-FieldMask": "*",
-          "X-Goog-Api-Key": `${process.env.EXPO_PUBLIC_GOOGLE_API_KEY}`
-        },
-        body: JSON.stringify({
-          textQuery: "bomb shelter around me",
-        }),
-      }
-    );
-    const {places} = await response.json();
-    setHref(places[0].googleMapsUri);
-  };
-  const openURL = async (url : string) => {
-    const supported = await Linking.canOpenURL(url);
-  
-    if (supported) {
-      await Linking.openURL(url);
-    } else {
-      console.log(`Don't know how to open this URL: ${url}`);
-    }
-  };
   return (
     <View style={styles.container}>
-      <Text style={styles.title}>Tab One</Text>
-      <TouchableOpacity onPress={fetchApi}>
-        <Text>Fetch</Text>
-      </TouchableOpacity>
-      <TouchableOpacity onPress={() => openURL(href)}>
-        <Text>Open</Text>
-      </TouchableOpacity>
+      <Text style={styles.title}>ShelterGo</Text>
       <View
         style={styles.separator}
         lightColor="#eee"
@@ -56,11 +22,15 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
     alignItems: "center",
-    justifyContent: "center",
+    paddingTop:40,
+    backgroundColor:"#569de8"
   },
   title: {
     fontSize: 20,
     fontWeight: "bold",
+    color:"#569de8",
+    backgroundColor:"#e8f3ff",
+    padding:10,
   },
   separator: {
     marginVertical: 30,
